ui: minor code refactor - moved utils.js to utils folder
